Add a Tor test template and several experimental tests.
tor_exit_ip_check fetches the egress IP by way of IP reflector service
and compares with the advertised ORPort

tor_http_requests fetches a set of URLs over each exit in the consensus.

tor_tls_handshake performs a TLS handshake to a set of endpoints for
each exit in the consensus. Unfortunately, python openssl bindings are
rather broken and this test does not correctly obtain the negotiated
cipher suite.